+++ common/src/main/java/whocraft/tardis_refined/client/TRKeybinds.java
@@ -1,0 +1,12 @@
+package whocraft.tardis_refined.client;
+
+import net.minecraft.client.KeyMapping;
+import org.lwjgl.glfw.GLFW;
+import whocraft.tardis_refined.TardisRefined;
+
+public class TRKeybinds {
+
+    public static KeyMapping EXIT_EXTERIOR_VIEW = new KeyMapping("exit_exterior_view", GLFW.GLFW_KEY_TAB, TardisRefined.NAME);
+
+
+}