+++ common/src/main/java/whocraft/tardis_refined/common/capability/player/TardisPilot.java
@@ -2,6 +2,7 @@ package whocraft.tardis_refined.common.capability.player;
 
 import net.minecraft.nbt.CompoundTag;
 import net.minecraft.server.level.ServerPlayer;
+import net.minecraft.world.entity.player.Abilities;
 import net.minecraft.world.entity.player.Player;
 import whocraft.tardis_refined.common.capability.tardis.TardisLevelOperator;
 import whocraft.tardis_refined.common.tardis.TardisNavLocation;
@@ -10,6 +11,8 @@ import javax.annotation.Nullable;
 import java.util.UUID;
 
 public interface TardisPilot {
+
+    void updatePlayerAbilities(ServerPlayer player, Abilities abilities, boolean isWatcher);
 
     void setupPlayerForInspection(ServerPlayer serverPlayer, TardisLevelOperator tardisLevelOperator, TardisNavLocation spectateTarget);