+++ common/src/main/java/whocraft/tardis_refined/common/mixin/PlayerRenderMixin.java
@@ -1,6 +1,9 @@
 package whocraft.tardis_refined.common.mixin;
 
+import com.mojang.blaze3d.vertex.PoseStack;
 import net.minecraft.client.Minecraft;
+import net.minecraft.client.player.AbstractClientPlayer;
+import net.minecraft.client.renderer.MultiBufferSource;
 import net.minecraft.client.renderer.entity.player.PlayerRenderer;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
@@ -14,10 +17,10 @@ import java.util.Objects;
 public class PlayerRenderMixin {
 
     @Inject(method = "render(Lnet/minecraft/client/player/AbstractClientPlayer;FFLcom/mojang/blaze3d/vertex/PoseStack;Lnet/minecraft/client/renderer/MultiBufferSource;I)V", at = @At("HEAD"), cancellable = true)
-    private void tick(CallbackInfo ci) {
+    private void render(AbstractClientPlayer abstractClientPlayer, float f, float g, PoseStack poseStack, MultiBufferSource multiBufferSource, int i, CallbackInfo ci) {
 
-       TardisPlayerInfo.get(Minecraft.getInstance().player).ifPresent(tardisPlayerInfo -> {
-           if (tardisPlayerInfo.isViewingTardis() && Objects.equals(Minecraft.getInstance().player.getStringUUID(), tardisPlayerInfo.getPlayer().getStringUUID())) {
+       TardisPlayerInfo.get(abstractClientPlayer).ifPresent(tardisPlayerInfo -> {
+           if (tardisPlayerInfo.isViewingTardis()) {
                ci.cancel();
            }
        });