+++ forge/src/main/java/whocraft/tardis_refined/common/data/LangProviderEnglish.java
@@ -258,6 +258,7 @@ public class LangProviderEnglish extends LanguageProvider {
         /*Overlay Messages*/
         add(ModMessages.ASCEND_KEY, "Ascend: %s");
         add(ModMessages.DESCEND_KEY, "Descend: %s");
+        add(ModMessages.EXIT_EXTERNAL_VIEW, "Exit External View: %s");
 
         /*Upgrades*/
         addUpgrade(TRUpgrades.CHAMELEON_CIRCUIT_SYSTEM.get(), "Chameleon Circuit", "Allows the TARDIS to change it's shape");